requirements.txt
# streamlit>=1.28.0
# boto3>=1.34.0
# networkx>=3.1.0
# pyvis>=0.3.2
# cryptography>=41.0.0
# python-dotenv>=1.0.0
# urllib3>=2.0.0

# IAM-X-Ray Core Dependencies (Beta v0.1.0)
# Runtime essentials only; pinned for reproducibility (Nov 2025)

streamlit==1.51.0
boto3==1.41.2
cryptography==46.0.3
networkx==3.5
pyvis==0.3.2
python-dotenv==1.2.1

# CI/Testing (from .github/workflows/ci.yml)
pytest==9.0.1
pytest-cov==7.0.0